+++ remix-debug/package.json
@@ -28,7 +28,6 @@
     "npm-run-all": "^4.1.2",
     "fast-async": "^6.1.2",
     "remix-lib": "^0.2.9",
-    "remix-solidity": "^0.1.11",
     "solc": "https://github.com/ethereum/solc-js"
   },
   "devDependencies": {
