+++ remix-debug/src/decoder/types/Enum.js
