+++ remix-solidity/test/tests.js
@@ -1,7 +1,3 @@
-require('./decoder/decodeInfo.js')
-require('./decoder/storageLocation.js')
-require('./decoder/storageDecoder.js')
-require('./decoder/localDecoder.js')
 
 require('./analysis/staticAnalysisCommon-test.js')
 require('./analysis/staticAnalysisIntegration-test.js')
